requirements.txt
discord.py>=2.3.2
httpx>=0.27.0
python-dotenv>=1.0.0
aiohttp>=3.9.0